highp -> lowp for demo

+++ js/game.js
@@ -20,7 +20,7 @@
 		{
 			antialias: true,
 			stencil: false,
-			precision: "highp",
+			precision: "lowp",
 			preserveDrawingBuffer: false,
 			depth: true
 		});
